+++ include/aligator/modelling/multibody/constrained-rnea.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
-#include "./fwd.hpp"
-#include "aligator/fwd.hpp"
+#include "aligator/modelling/multibody/fwd.hpp"
 
 #include <pinocchio/algorithm/compute-all-terms.hpp>
 #include <pinocchio/algorithm/contact-jacobian.hpp>
@@ -61,8 +60,16 @@ void underactuatedConstrainedInverseDynamics(
   res = qr.solve(nle);
 }
 
-} // namespace aligator
-
 #ifdef ALIGATOR_ENABLE_TEMPLATE_INSTANTIATION
-#include "./constrained-rnea.txx"
+extern template void underactuatedConstrainedInverseDynamics<
+    context::Scalar, context::ConstVectorRef, context::ConstVectorRef,
+    context::ConstMatrixRef, context::VectorRef, context::Options>(
+    const context::PinModel &, context::PinData &,
+    const Eigen::MatrixBase<context::ConstVectorRef> &,
+    const Eigen::MatrixBase<context::ConstVectorRef> &,
+    const Eigen::MatrixBase<context::ConstMatrixRef> &,
+    const StdVectorEigenAligned<context::RCM> &,
+    StdVectorEigenAligned<context::RCD> &,
+    const Eigen::MatrixBase<context::VectorRef> &);
 #endif
+} // namespace aligator