+++ include/aligator/modelling/multibody/fwd.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "aligator/context.hpp"
 #include <pinocchio/algorithm/frames.hpp>
 
 namespace aligator {
@@ -11,5 +12,18 @@ struct frame_api {
 protected:
   pinocchio::FrameIndex pin_frame_id_;
 };
+
+template <typename> struct MultibodyConfiguration;
+template <typename> struct MultibodyPhaseSpace;
+
+namespace context {
+using PinModel = pinocchio::ModelTpl<Scalar, Options>;
+using PinData = pinocchio::DataTpl<Scalar, Options>;
+
+using RCM = pinocchio::RigidConstraintModelTpl<Scalar, Options>;
+using RCD = pinocchio::RigidConstraintDataTpl<Scalar, Options>;
+using MultibodyConfiguration = MultibodyConfiguration<Scalar>;
+using MultibodyPhaseSpace = MultibodyPhaseSpace<Scalar>;
+} // namespace context
 
 } // namespace aligator